requirements.txt
python-binance
python-dotenv
pandas
ta